Add an availability MQTT topic and set up a will
This commit basically does 3 things:

 - Publish "<workgroup>/<machineId>/status" with the value "online"
   when connected.

 - Set up a will topic, so that whenever the ANAVI Thermometer is
   disconnected from the MQTT broker, the broker will publish
   "<workgroup>/<machineId>/status" with the value "offline"

 - Add "availability_topic" to the MQTT Discovery message, so that
   Home Assistant will listen to "<workgroup>/<machineId>/status" to
   determine if the ANAVI Thermometer is available or not.

Note that all the sensors of a particular ANAVI Thermometer will share
the same status topic, so they will all be either online or offline.
In a way, it would be better to have separate availability topics for
each sensor, but MQTT only supports a single will topic, so only one
of the sensors could be set "offline" automatically.