De-publish sensor values when the DHT22 isn't available
Publish an empty string when the DHT22 sensor refuses to publish a
value.  This will make the value disappar from the MQTT broker.

Unfortunately, it seems that Home Assistant will retain the last seen
value, so this is perhaps not the best thing to do.  It would be nice
if Home Assistant supported two availability topics for a sensor --
that way we could have one topic that is set to "offline" from an MQTT
will to signal that the entire ANAVI Thermometer is offline, and
another topic that the ANAVI Thermometer could use to indicate that a
particular sensor is offline.

The next commit has an alternative solution...