Use a separate availability topic for DHT22-derived values
The new "<workgroup>/<machineId>/dht22status" topic will be set to
"online" if the DHT22 sensor is working properly, and "offline" if we
can't read anything from it.  The /air/temperature and /air/humidity
devices will use the dht22status topic as the availability topic.

There is one drawback with this change: there is no MQTT will that
automatically sets dht22status to offline when the connection to the
ANAVI Thermometer drops.  However, a simple Home Assistant automation
can be created that automatically sets dht22status to offline whenever
status is set to offline.  For example, if the the workgroup is
"workgroup" and the machineId is "deadbeef", the following automation
would to the job:

- id: '4711'
  alias: Track ANAVI Thermometer disconnect in dht22status
  description: ''
  trigger:
  - topic: workgroup/deadbeef/status
    payload: offline
    platform: mqtt
  condition: []
  action:
  - service: mqtt.publish
    data:
      topic: workgroup/deadbeef/dht22status
      payload: offline
      qos: 0
      retain: true

Don't publish null values when the sensor is unavailable.