Use one MQTT connection and one status topic per sensor
Having multiple status topics means that we can immediately inform
Home Assistant when individual sensors become available or
unavailable.  For instance, the DS18B20 sensor can now be hotplugged
and "online" or "offline" messages will be sent to the
workgroup/machineId/status/ds18b20 MQTT topic to indicate if the
sensor is available or not.

Having multiple MQTT connections means we can have a will for each of
the +/+/status/+ topics, so that they will all receive an "offline"
message when the ANAVI Thermometer loses connection to the MQTT
broker.

Connections will only be created for sensors that are detected.  The
special "esp8266" connection is used for the board itself.

To avoid race conditions between the status messages and the sensor
readings, each update is sent via the connection that belongs to to
sensor that originates the message.

This commit also makes a few other changes:

 - Send MQTT Discovery messages only sent when the connection is
   established.  The only exceptions are the bmp180-altitude and
   bmp180-slp sensors.  Since they are only published when the
   sea-level pressure and altitude are configured via MQTT messages,
   they are still sent from the publishState() function.

 - Add hot-plug support for the DS18B20 sensor.